refactor: ~src out and interfaces
