+++ src/pages/Home/index.tsx
@@ -1,7 +1,7 @@
 import { useQuery } from '@tanstack/react-query'
 import React, { useEffect, useMemo } from 'react'
 import { useSearchParams, createSearchParams } from 'react-router-dom'
-import { fetchItems } from '~src/utils/fetchItems'
+import { fetchItems } from 'utils/fetchItems'
 import Header from './Header'
 import styled from 'styled-components'
 import RegistryDetails from './RegistryDetails'
@@ -9,9 +9,9 @@ import SubmitEntries from './SubmitEntries'
 import Search from './Search'
 import LoadingItems from './LoadingItems'
 import EntriesList from './EntriesList'
-import Footer from '~src/components/Footer'
+import Footer from 'components/Footer'
 import Pagination from './Pagination'
-import { fetchItemCounts } from '~src/utils/itemCounts'
+import { fetchItemCounts } from 'utils/itemCounts'
 import DetailsModal from './DetailsModal'
 import Filters from './Filters'
